ENH: Move compare_lr_test to LikelihoodModelResults
See #2436
Closes #2440
